+++ Tests/Integration/MSFT_xNetAdapterName_nameonly.config.ps1
@@ -1,0 +1,10 @@
+configuration MSFT_xNetAdapterName_Config_NameOnly {
+    Import-DscResource -ModuleName xNetworking
+
+    node localhost {
+        xNetAdapterName Integration_Test {
+            Name                 = $Node.Name
+            NewName              = $Node.NewName
+        }
+    }
+}